Use POST instead of GET request for OmniAuth call
The app was working in production, but not on development. Changing the
OAuth request to be a POST via form button fixed things locally.

Possibly related to this: https://github.com/omniauth/omniauth/wiki/Resolving-CVE-2015-9284#using-get-requests

Not sure why I hit this error without upgrading, but it looks like this
would be needed eventually: https://github.com/omniauth/omniauth/wiki/Upgrading-to-2.0#racksinatra